Use int handles

+++ osu.Framework/Audio/Handles/SafeBassChannelHandle.cs
@@ -1,14 +1,13 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
 using ManagedBass;
 
 namespace osu.Framework.Audio.Handles
 {
     public class SafeBassChannelHandle : SafeBassHandle
     {
-        public SafeBassChannelHandle(IntPtr handle, bool ownsHandle)
+        public SafeBassChannelHandle(int handle, bool ownsHandle)
             : base(handle, ownsHandle)
         {
         }
