+++ osu.Framework/Audio/Handles/SafeBassSampleHandle.cs
@@ -1,14 +1,13 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
 using ManagedBass;
 
 namespace osu.Framework.Audio.Handles
 {
     public sealed class SafeBassSampleHandle : SafeBassHandle
     {
-        public SafeBassSampleHandle(IntPtr handle, bool ownsHandle)
+        public SafeBassSampleHandle(int handle, bool ownsHandle)
             : base(handle, ownsHandle)
         {
         }
